requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
Pillow==10.2.0
aiofiles==23.2.1
